Revert "Reverts "Remove Cocoon from `dev/devicelab`, keeping Skia perf stats upload. (#165749)" (#165754)"
This reverts commit 0a2d9f56585824689d26a751f3326bf22865ea11.
