+++ dev/devicelab/bin/test_runner.dart
@@ -8,12 +8,12 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 
 import 'package:flutter_devicelab/command/test.dart';
-import 'package:flutter_devicelab/command/upload_results.dart';
+import 'package:flutter_devicelab/command/upload_metrics.dart';
 
 final CommandRunner<void> runner =
     CommandRunner<void>('devicelab_runner', 'DeviceLab test runner for recording test results')
       ..addCommand(TestCommand())
-      ..addCommand(UploadResultsCommand());
+      ..addCommand(UploadMetricsCommand());
 
 Future<void> main(List<String> rawArgs) async {
   unawaited(
